refactor to use es6 export
