+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules
-bower_components
 .npmrc
 
 npm-debug.log
+yarn-error.log
