--- .jscsrc
@@ -1,83 +1,0 @@
-{
-    "requireCurlyBraces": [
-        "if",
-        "else",
-        "for",
-        "while",
-        "do",
-        "try",
-        "catch"
-    ],
-    "requireOperatorBeforeLineBreak": true,
-    "maximumLineLength": {
-        "value": 150,
-        "allowComments": true,
-        "allowRegex": true
-    },
-    "validateIndentation": 4,
-    "validateQuoteMarks": "'",
-    "disallowMultipleLineStrings": true,
-    "disallowMixedSpacesAndTabs": true,
-    "disallowTrailingWhitespace": true,
-    "disallowSpaceAfterPrefixUnaryOperators": true,
-    "requireSpaceAfterKeywords": [
-        "if",
-        "else",
-        "for",
-        "while",
-        "do",
-        "switch",
-        "return",
-        "try",
-        "catch"
-    ],
-    "requireSpaceBeforeBinaryOperators": [
-        "=",
-        "+=",
-        "-=",
-        "*=",
-        "/=",
-        "%=",
-        "<<=",
-        ">>=",
-        ">>>=",
-        "&=",
-        "|=",
-        "^=",
-        "+=",
-        "+",
-        "-",
-        "*",
-        "/",
-        "%",
-        "<<",
-        ">>",
-        ">>>",
-        "&",
-        "|",
-        "^",
-        "&&",
-        "||",
-        "===",
-        "==",
-        ">=",
-        "<=",
-        "<",
-        ">",
-        "!=",
-        "!=="
-    ],
-    "requireSpaceAfterBinaryOperators": true,
-    "requireSpacesInConditionalExpression": true,
-    "requireSpaceBeforeBlockStatements": true,
-    "requireSpacesInForStatement": true,
-    "requireLineFeedAtFileEnd": true,
-    "requireSpacesInFunctionExpression": {
-       "beforeOpeningRoundBrace": true,
-       "beforeOpeningCurlyBrace": true
-    },
-    "disallowSpacesInsideObjectBrackets": "all",
-    "disallowSpacesInsideArrayBrackets": "all",
-    "disallowSpacesInsideParentheses": true,
-    "disallowMultipleLineBreaks": true
-}
