--- .jshintrc
@@ -1,29 +1,0 @@
-{
-    "browser": true,
-    "camelcase": true,
-    "curly": true,
-    "eqeqeq": true,
-    "evil": false,
-    "immed": true,
-    "latedef": false,
-    "loopfunc": true,
-    "maxlen": 150,
-    "multistr": true,
-    "newcap": true,
-    "noarg": true,
-    "noempty": true,
-    "nonbsp": true,
-    "nonew": true,
-    "quotmark": "single",
-    "sub": true,
-    "undef": true,
-    "unused": "vars",
-    "globals" : {
-        "__dirname": false,
-        "d3" : false,
-        "dc" : true,
-        "module": false,
-        "process": false,
-        "require": false
-    }
-}
