--- grunt/options/clean.js
@@ -1,11 +1,0 @@
-module.exports = {
-    build: {
-        files: [{
-            dot: true,
-            src: [
-                '.tmp',
-                '<%= config.dist %>/**/*'
-            ]
-        }]
-    }
-};
