--- grunt/options/jscs.js
@@ -1,16 +1,0 @@
-module.exports = {
-    options: {
-        config: '.jscsrc'
-    },
-    src: {
-        src: [
-            '<%= config.src %>/**/*.js',
-        ]
-    },
-    grunt: {
-        src: [
-            'Gruntfile.js',
-            '<%= config.grunt %>/**/*.js'
-        ]
-    }
-};
