--- grunt/options/jshint.js
@@ -1,12 +1,0 @@
-module.exports = {
-    options: {
-        jshintrc: '.jshintrc',
-    },
-    src: [
-        '<%= config.src %>/**/*.js',
-    ],
-    grunt: [
-        'Gruntfile.js',
-        '<%= config.grunt %>/**/*.js'
-    ]
-};
