--- grunt/options/notify.js
@@ -1,26 +1,0 @@
-module.exports = {
-    watch: {
-        options: {
-            title: 'Grunt Watch Task',
-            message: 'Task completed successfully'
-        }
-    },
-    serve: {
-        options: {
-            title: 'Grunt Serve Task',
-            message: 'Task completed successfully - start watching'
-        }
-    },
-    build: {
-        options: {
-            title: 'Grunt Build Task',
-            message: 'Started'
-        }
-    },
-    buildComplete: {
-        options: {
-            title: 'Grunt Build Task',
-            message: 'Task completed successfully'
-        }
-    },
-};
