--- grunt/options/uglify.js
@@ -1,16 +1,0 @@
-module.exports = {
-    options: {
-        sourceMap: false,
-        screwIE8: true,
-        preserveComments: 'some'
-    },
-    build: {
-        files: [{
-            expand: true,
-            cwd: '<%= config.dist %>/',
-            src: '**/*.js',
-            dest: '<%= config.dist %>/',
-            ext: '.min.js'
-        }]
-    }
-};
