--- grunt/test.js
@@ -1,8 +1,0 @@
-module.exports = function (grunt) {
-    grunt.registerTask('test', [
-        'jshint:src',
-        'jshint:grunt',
-        'jscs:src',
-        'jscs:grunt'
-    ]);
-};
