+++ package.json
@@ -1,23 +1,17 @@
 {
   "name": "dc-addons-bubble-chart",
-  "version": "1.0.1",
-  "main": "./dist/dc-bubble-cloud.js",
-  "dependencies": {},
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/eileenprice/dc-addons-bubble-chart.git"
+  "version": "2.0.0",
+  "repository": "https://github.com/eileenprice/dc-addons-bubble-chart",
+  "dependencies": {
+    "d3": "^5.7.0",
+    "dc": "^3.0.8"
   },
   "devDependencies": {
-    "glob": "^7.0.0",
-    "grunt": "^1.0.0",
-    "grunt-contrib-clean": "^1.0.0",
-    "grunt-contrib-concat": "^1.0.0",
-    "grunt-contrib-jshint": "^1.0.0",
-    "grunt-contrib-uglify": "^1.0.0",
-    "grunt-jscs": "^2.0.0",
-    "grunt-notify": "^0.4.1",
-    "load-grunt-tasks": "^3.3.0",
-    "time-grunt": "^1.2.0"
+    "eslint": "^5.7.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^4.0.1",
+    "eslint-plugin-standard": "^4.0.0"
   },
   "engines": {
     "node": ">=6"
